requirements.txt
streamlit==1.28.0
flask==2.3.3
flask-socketio==5.3.6
python-socketio==5.9.0
websockets==11.0.3
pyaudio==0.2.11
numpy==1.24.3
scipy==1.11.2
wave==0.0.2
requests==2.31.0
python-dotenv==1.0.0
openai-whisper==20230918
torch==2.0.1
torchaudio==2.0.2
groq
pycaw==20230407  # For Windows audio control
sounddevice==0.4.6  # Alternative audio library
